// interfaces/post.ts
type PostType = {
  key: number
  client: string
  tags: string
  title: string
  date: string
  topImage: string
  slug: string
  btn: {
    label: string,
    url: string
    },
    btn2: {
      label: string,
      url: string
      },
      btn3: {
        label: string,
        url: string
        },
        btn4: {
          label: string,
          url: string
          },
  content: string
}

export default PostType